+++ internal/api/weatherapi/openmeteo/mocks/mocks.go
